Move superblock validation to the quorum
The quorum orchestrates superblock production and intake. It's a more
appropriate location for the superblock validation code. The original
source file contains low-level superblock data structures that do not
provide application services.
